+++ src/actions/internal.ts
@@ -2,3 +2,4 @@
 export * from "./Action";
 export * from "./CrudAction";
 export * from "./RootAction";
+export * from "./LoginAction";